+++ src/SwedbankPay.Sdk.Infrastructure/PaymentInstruments/Invoice/InvoicePaymentDto.cs
@@ -4,11 +4,11 @@ namespace SwedbankPay.Sdk.PaymentInstruments.Invoice
 {
     internal class InvoicePaymentDto
     {
-        public int Amount { get; set; }
-        public int VatAmount { get; set; }
-        public int RemainingCaptureAmount { get; set; }
-        public int RemainingCancellationAmount { get; set; }
-        public int RemainingReversalAmount { get; set; }
+        public long Amount { get; set; }
+        public long VatAmount { get; set; }
+        public long RemainingCaptureAmount { get; set; }
+        public long RemainingCancellationAmount { get; set; }
+        public long RemainingReversalAmount { get; set; }
         public InvoicePaymentAuthorizationListDto Authorizations { get; set; }
         public CancellationListResponseDto Cancellations { get; set; }
         public CaptureListResponseDto Captures { get; set; }
