+++ src/SwedbankPay.Sdk.Infrastructure/PaymentOrders/OrderItemDto.cs
@@ -15,8 +15,8 @@
         public int UnitPrice { get; set; }
         public int DiscountPrice { get; set; }
         public int VatPercent { get; set; }
-        public int Amount { get; set; }
-        public int VatAmount { get; set; }
+        public long Amount { get; set; }
+        public long VatAmount { get; set; }
 
         internal IOrderItem Map()
         {
